Changed float to double

+++ include/computations.hpp
@@ -5,14 +5,14 @@
 
 class Computations {
     public:
-        float t;
+        double t;
         Sphere object;
         Tuple point;
         Tuple eyev;
         Tuple normalv;
         bool inside;
 
-        Computations(float t, Sphere object, Tuple point, Tuple eyev, Tuple normalv);
+        Computations(double t, Sphere object, Tuple point, Tuple eyev, Tuple normalv);
 };
 
 Computations prepare_computations(Intersection i, Ray r);
